feat(ActivityView): add styles and refactor feature flag selector for activity redesign
