+++ app/components/Views/ActivityView/selectors/featureFlags/index.ts
@@ -1,5 +1,5 @@
 import { createSelector } from 'reselect';
-import { selectRemoteFeatureFlags } from '..';
+import { selectRemoteFeatureFlags } from '../../../../../selectors/featureFlagController';
 
 export const selectIsActivityRedesignEnabled = createSelector(
   selectRemoteFeatureFlags,
